AddCompareLocaleActivityの内容をSearchListFragmentに移した
