+++ app/src/main/java/honkot/gscheduler/di/AppComponent.java
@@ -6,6 +6,7 @@ import dagger.Component;
 import honkot.gscheduler.AddCompareLocaleActivity;
 import honkot.gscheduler.MainActivity;
 import honkot.gscheduler.fragment.RecordListFragment;
+import honkot.gscheduler.fragment.SearchListFragment;
 
 @Singleton
 @Component(modules = AppModule.class)
@@ -14,6 +15,8 @@ public interface AppComponent {
     void inject(MainActivity mainActivity);
 
     void inject(RecordListFragment fragment);
+
+    void inject(SearchListFragment fragment);
 
     void inject(AddCompareLocaleActivity activity);
 }
