+++ app/src/main/res/layout/fragment_search_list.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android">
+    <LinearLayout
+        android:orientation="vertical"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+        <SearchView
+            android:id="@+id/searchView"
+            android:background="@drawable/search_bg"
+            android:layout_margin="@dimen/searchView_margin"
+            android:iconifiedByDefault="false"
+            android:queryHint="@string/search"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
+        <ListView
+            android:id="@+id/listView"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:fastScrollEnabled="true"
+            android:fastScrollAlwaysVisible="true" />
+    </LinearLayout>
+</layout>
